+++ src/viur/datastore/config.py
@@ -1,0 +1,6 @@
+"""
+	This file just holds some configuration variables that will influence
+	the behaviour of this library.
+"""
+traceQueries = False  # If set, we'll log each query we run
+SkeletonInstanceRef = None  # A reference to the skeleton container of ViUR. Unless set, fetch() and getSkel() will fail